+++ packages/article/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.4.46](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.45...@times-components/article@7.4.46) (2020-02-07)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.4.45](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.44...@times-components/article@7.4.45) (2020-02-06)